streamlit-ML-CC-anormally/requirements.txt
# Core dependencies
streamlit==1.30.0
pandas==1.5.3

# Hugging Face Transformers & Torch
transformers==4.37.0


# AutoTrain Advanced
autotrain-advanced==0.6.0

# Additional dependencies for stability
scikit-learn==1.3.0
datasets==2.17.0
sentencepiece==0.1.99
protobuf==4.24.0
